chore(worker_bee): Refactor tests
Refactors worker_bee tests to allow for DDC/dart2js testing

commit-id:87173490
