+++ packages/worker_bee/e2e/.gitignore
@@ -1,0 +1,12 @@
+# Files and directories created by pub.
+.dart_tool/
+.packages
+pubspec.lock
+
+# Conventional directory for build output.
+build/
+
+# JS outputs
+*.js
+*.js.map
+*.js.deps
