+++ packages/worker_bee/e2e/README.md
@@ -1,0 +1,3 @@
+# Worker Bee E2E Tests
+
+E2E tests for the [`worker_bee`](../worker_bee) package.
