+++ packages/worker_bee/e2e/lib/e2e_worker.worker.js.dart
@@ -1,7 +1,7 @@
 // Generated by worker_bee_builder.
 
 import 'package:aws_common/aws_common.dart';
-import 'package:e2e_test/e2e_worker.dart';
+import 'package:e2e/e2e_worker.dart';
 
 /// The JS implementation of [E2EWorker].
 class E2EWorkerImpl extends E2EWorker {
@@ -9,6 +9,6 @@ class E2EWorkerImpl extends E2EWorker {
   String get name => 'E2EWorker';
   @override
   List<String> get fallbackUrls => zDebugMode
-      ? const ['packages/e2e_test/workers.debug.dart.js']
-      : const ['packages/e2e_test/workers.release.dart.js'];
+      ? const ['packages/e2e/workers.debug.dart.js']
+      : const ['packages/e2e/workers.release.dart.js'];
 }
