+++ packages/worker_bee/e2e/lib/e2e_worker_no_result.worker.js.dart
@@ -1,7 +1,7 @@
 // Generated by worker_bee_builder.
 
 import 'package:aws_common/aws_common.dart';
-import 'package:e2e_test/e2e_worker_no_result.dart';
+import 'package:e2e/e2e_worker_no_result.dart';
 
 /// The JS implementation of [E2EWorkerNoResult].
 class E2EWorkerNoResultImpl extends E2EWorkerNoResult {
@@ -11,16 +11,16 @@ class E2EWorkerNoResultImpl extends E2EWorkerNoResult {
   String get jsEntrypoint {
     // Flutter web release builds must use the bundled asset.
     if (zIsFlutter && !zDebugMode) {
-      return 'assets/packages/e2e_test/lib/workers.min.js';
+      return 'assets/packages/e2e/lib/workers.min.js';
     }
     // Default to the compiled, published worker.
     return zDebugMode
-        ? 'packages/e2e_test/workers.js'
-        : 'packages/e2e_test/workers.min.js';
+        ? 'packages/e2e/workers.js'
+        : 'packages/e2e/workers.min.js';
   }
 
   @override
   List<String> get fallbackUrls => zDebugMode
-      ? const ['packages/e2e_test/workers.debug.dart.js']
-      : const ['packages/e2e_test/workers.release.dart.js'];
+      ? const ['packages/e2e/workers.debug.dart.js']
+      : const ['packages/e2e/workers.release.dart.js'];
 }
