+++ packages/worker_bee/e2e/lib/workers.dart
@@ -20,9 +20,10 @@
   E2EWorkerThrows,
 ])
 
-import 'package:e2e_test/e2e_worker.dart';
-import 'package:e2e_test/e2e_worker_no_result.dart';
-import 'package:e2e_test/e2e_worker_null_result.dart';
-import 'package:e2e_test/e2e_worker_throws.dart';
-import 'package:e2e_test/e2e_worker_void_result.dart';
 import 'package:worker_bee/worker_bee.dart';
+
+import 'e2e_worker.dart';
+import 'e2e_worker_no_result.dart';
+import 'e2e_worker_null_result.dart';
+import 'e2e_worker_throws.dart';
+import 'e2e_worker_void_result.dart';
