+++ packages/worker_bee/e2e/lib/workers.release.dart
@@ -1,10 +1,10 @@
 // Generated by worker_bee_builder.
 
-import 'package:e2e_test/e2e_worker.dart';
-import 'package:e2e_test/e2e_worker_no_result.dart';
-import 'package:e2e_test/e2e_worker_null_result.dart';
-import 'package:e2e_test/e2e_worker_throws.dart';
-import 'package:e2e_test/e2e_worker_void_result.dart';
+import 'package:e2e/e2e_worker.dart';
+import 'package:e2e/e2e_worker_no_result.dart';
+import 'package:e2e/e2e_worker_null_result.dart';
+import 'package:e2e/e2e_worker_throws.dart';
+import 'package:e2e/e2e_worker_void_result.dart';
 import 'package:worker_bee/worker_bee.dart';
 
 final _workers = <String, WorkerBeeBuilder>{
