+++ packages/worker_bee/e2e/mono_pkg.yaml
@@ -1,0 +1,13 @@
+sdk:
+  - stable
+  - dev
+
+stages:
+  - analyze_and_format:
+      - group:
+          - format
+          - analyze: --fatal-infos .
+  - build_test:
+      - test: --tags=build
+        sdk:
+          - stable
