+++ packages/worker_bee/e2e/pubspec.yaml
@@ -1,9 +1,20 @@
-name: e2e_test
+name: e2e
 description: E2E tests for the worker_bee package.
 publish_to: none
 
 environment:
   sdk: '>=2.17.0 <3.0.0'
+
+dependencies: 
+  aws_common:
+    path: ../../aws_common
+  built_collection: ^5.0.0
+  built_value: ^8.0.0
+  fixnum: ^1.0.0
+  meta: ^1.7.0
+  test: ^1.19.0
+  worker_bee:
+    path: ../worker_bee
 
 dependency_overrides:
   aws_common:
@@ -14,19 +25,9 @@ dependency_overrides:
 dev_dependencies:
   amplify_lints:
     path: ../../amplify_lints
-  aws_common:
-    path: ../../aws_common
   build_runner: ^2.0.0
-  build_test: ^2.0.0
+  build_verify: ^3.0.0
   build_web_compilers: ^3.2.3
-  built_collection: ^5.0.0
-  built_value: ^8.0.0
-  built_value_generator: 8.4.0
-  e2e:
-    path: ../e2e
-  meta: ^1.7.0
-  test: ^1.19.0
-  worker_bee:
-    path: ../worker_bee
+  built_value_generator: 8.3.3
   worker_bee_builder:
     path: ../worker_bee_builder
