+++ packages/worker_bee/e2e_flutter_test/pubspec.yaml
@@ -7,10 +7,16 @@ environment:
   sdk: ">=2.17.0 <3.0.0"
 
 dependencies:
-  e2e_test:
-    path: ../e2e_test
+  e2e:
+    path: ../e2e
   flutter:
     sdk: flutter
+
+dependency_overrides:
+  aws_common:
+    path: ../../aws_common
+  worker_bee:
+    path: ../worker_bee
 
 dev_dependencies:
   amplify_lints:
