+++ packages/worker_bee/e2e_test/test/e2e_js_test.dart
@@ -15,7 +15,7 @@
 @TestOn('browser')
 
 import 'package:aws_common/aws_common.dart';
-import 'package:e2e_test/common.dart';
+import 'package:e2e/common.dart';
 import 'package:test/test.dart';
 
 void main() {
@@ -26,14 +26,14 @@ void main() {
       test(
         'dart2Js',
         () => testWorker(
-          jsEntrypoint: 'packages/e2e_test/workers.js',
+          jsEntrypoint: 'packages/e2e/workers.js',
         ),
       );
 
       test(
         'dart2Js (m, O4)',
         () => testWorker(
-          jsEntrypoint: 'packages/e2e_test/workers.min.js',
+          jsEntrypoint: 'packages/e2e/workers.min.js',
         ),
       );
     });
@@ -42,14 +42,14 @@ void main() {
       test(
         'dart2Js',
         () => testWorkerNoResult(
-          jsEntrypoint: 'packages/e2e_test/workers.js',
+          jsEntrypoint: 'packages/e2e/workers.js',
         ),
       );
 
       test(
         'dart2Js (m, O4)',
         () => testWorkerNoResult(
-          jsEntrypoint: 'packages/e2e_test/workers.min.js',
+          jsEntrypoint: 'packages/e2e/workers.min.js',
         ),
       );
     });
@@ -58,14 +58,14 @@ void main() {
       test(
         'dart2Js',
         () => testWorkerVoidResult(
-          jsEntrypoint: 'packages/e2e_test/workers.js',
+          jsEntrypoint: 'packages/e2e/workers.js',
         ),
       );
 
       test(
         'dart2Js (m, O4)',
         () => testWorkerVoidResult(
-          jsEntrypoint: 'packages/e2e_test/workers.min.js',
+          jsEntrypoint: 'packages/e2e/workers.min.js',
         ),
       );
     });
@@ -74,14 +74,14 @@ void main() {
       test(
         'dart2Js',
         () => testWorkerNullResult(
-          jsEntrypoint: 'packages/e2e_test/workers.js',
+          jsEntrypoint: 'packages/e2e/workers.js',
         ),
       );
 
       test(
         'dart2Js (m, O4)',
         () => testWorkerNullResult(
-          jsEntrypoint: 'packages/e2e_test/workers.min.js',
+          jsEntrypoint: 'packages/e2e/workers.min.js',
         ),
       );
     });
@@ -90,14 +90,14 @@ void main() {
       test(
         'dart2Js',
         () => testWorkerThrows(
-          jsEntrypoint: 'packages/e2e_test/workers.js',
+          jsEntrypoint: 'packages/e2e/workers.js',
         ),
       );
 
       test(
         'dart2Js (m, O4)',
         () => testWorkerThrows(
-          jsEntrypoint: 'packages/e2e_test/workers.min.js',
+          jsEntrypoint: 'packages/e2e/workers.min.js',
         ),
       );
     });
