+++ packages/worker_bee/e2e_test/test/preamble_test.dart
@@ -45,11 +45,11 @@ void main() {
 
   group('preamble', () {
     test('bad assignment', () {
-      badAssignmentTest('packages/e2e_test/no_workers.js');
+      badAssignmentTest('packages/e2e/no_workers.js');
     });
 
     test('bad assignment (m, O4)', () {
-      badAssignmentTest('packages/e2e_test/no_workers.min.js');
+      badAssignmentTest('packages/e2e/no_workers.min.js');
     });
   });
 }
